camera-service/requirements.txt
flask>=3.0.0
flask-cors>=5.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
flirimageextractor>=1.4.0
watchdog>=4.0.0
